Core changes (not compiling yet)

+++ kratos/utilities/embedded_skin_utility.cpp
@@ -234,7 +234,7 @@ namespace Kratos
             case GeometryData::KratosGeometryType::Kratos_Triangle3D3:
                 return Kratos::make_shared<Triangle3D3< Node<3> > >(rNewNodesArray);
             default:
-                KRATOS_ERROR << "Implement the skin generation for the intersection geometry type: " << rOriginGeometryType;
+                KRATOS_ERROR << "Implement the skin generation for the intersection geometry type: " << static_cast<int>(rOriginGeometryType);
         }
     }
